[fix] バグの修正

+++ src/main/java/com/github/tanokun/tanorpg/listener/DamageEventListener.java
@@ -154,8 +154,8 @@ public class DamageEventListener implements Listener {
             int at_lvl = customEntity.getLEVEL();
             int vi_lvl = victim.getLEVEL();
             double atk = DamageManager.getDamage(customEntity.getATK(), customEntity.getINT(), customEntity.getAGI());
-            int damage = DamageManager.getCompDamage(atk, victim.getStatus(StatusType.DEF).getLevel(), at_lvl, vi_lvl, e.getDamager());
-            victim.setHAS_HP(victim.getHAS_HP() - damage);
+            long damage = DamageManager.getCompDamage(atk, victim.getStatus(StatusType.DEF).getLevel(), at_lvl, vi_lvl, e.getDamager());
+            victim.setHAS_HP((int) (victim.getHAS_HP() - damage));
             if (victim.getHAS_HP() <= 0){
                 new BukkitRunnable(){
                     @Override
